Refactor to more OOP
