+++ filename.txt
@@ -1,5 +1,8 @@
-D|F||2021-10-10
-T|F|a
 D|T||2020-11-11
 D|F||2021-11-11
-D|F|abc |2021-11-11
+D|F||2021-11-11
+T|F|asa
+D|T|do it ok |2022-02-02
+E|F||y jd
+E|F||fjso /by da
+E|F|sadas |dasdas
