+++ src/main/java/Storage.java
@@ -1,4 +1,5 @@
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
@@ -19,5 +20,14 @@ public class Storage {
         }
         myReader.close();
         return result;
+    }
+
+    public static void writeToFile() throws IOException {
+        FileWriter myWriter = new FileWriter("filename.txt");
+        for (int i = 0; i < List.todos.size(); i++) {
+            myWriter.write(List.todos.get(i).toDataString());
+            myWriter.write(System.lineSeparator());
+        }
+        myWriter.close();
     }
 }
